% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotWeatherSummary.R
\name{aggRain}
\alias{aggRain}
\title{aggRain}
\usage{
aggRain(weatherInfo)
}
\arguments{
\item{weatherInfo}{weather columns from a dataframe}
}
\value{
Column of aggregate rain
}
\description{
This function returns combines all rain columns into an aggregate over all time
}

