% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotWeatherSummary.R
\name{format_for_mark}
\alias{format_for_mark}
\title{format_for_mark}
\usage{
format_for_mark(shot_weather)
}
\arguments{
\item{data.frame}{shot_weather weather and shots for tournament}

\item{boolean}{for_mark format data for mark or no}
}
\value{
Data Frame row of just weather infor for each shot
}
\description{
This function takes relevant columns from a shot_weather dataframe and turns it into a regular data frame
}
\examples{
format_for_mark(all.shot.weather)

}
\keyword{golf}
\keyword{pga}

