% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCleanProvizForCourse.R
\name{getCleanProvizForCourse}
\alias{getCleanProvizForCourse}
\title{Get Proviz in the cleaned format}
\usage{
getCleanProvizForCourse(courseName)
}
\arguments{
\item{event}{event returned}
}
\value{
Data Frame of events for the given year range.
}
\description{
This function returns a data frame of events for the pga tour between specific years, or at specific courses/tournaments
}
\examples{
getEventsForYears(2010, 2015)


}
\keyword{golf}
\keyword{pga}

