% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElevationAtPoints.R
\name{getElevationAtPoints}
\alias{getElevationAtPoints}
\title{Elevation for a string of points}
\usage{
getElevationAtPoints(pointString)
}
\arguments{
\item{string}{list of points}
}
\value{
vector of elevations in feet
}
\description{
This function finds elevation for a string of points, separated by |
}
\examples{
getElevationAtPoints("40.743484674348,-73.454984245498|40.744527774453,-73.454463145446|40.74670037467,-73.453455145346")
}

