% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocationForEvents.R
\name{getLocationForEvents}
\alias{getLocationForEvents}
\title{Get location for pga events}
\usage{
getLocationForEvents(evnts)
}
\arguments{
\item{dataframe}{evnts}
}
\value{
Data Frame of events with coordinates
}
\description{
This function takes a dataframe of events and returns a dataframe with lat and longituded 
binded to the original dataframe
}
\examples{
getLocationForEvents(events)


}
\keyword{golf}
\keyword{pga}

