% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocationForPGACourse.R
\name{getLocationForPGACourse}
\alias{getLocationForPGACourse}
\title{Get coordinates for pga course}
\usage{
getLocationForPGACourse(course)
}
\arguments{
\item{string}{course, name of course}
}
\value{
vector of c(lat, long)
}
\description{
This function finds coordinates for a course name
}
\examples{
getLocationForPGACourse("Baltusrol GC")
}

