% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPGAEvents.R
\name{getPGAEvents}
\alias{getPGAEvents}
\title{Get pga events}
\usage{
getPGAEvents(startYear = 0, endYear = 10000, course = NA,
  tournament = NA)
}
\arguments{
\item{integer}{input first year of events to return, Defaults to 0 (no lower bound)}

\item{integer}{input last year of events to return, Defaults to 10000 (no upper bound)}

\item{string}{course Name of course to filter by}

\item{string}{tournament Name of tournament to filter by}
}
\value{
Data Frame of events for the given year range.
}
\description{
This function returns a data frame of events for the pga tour between specific years, or at specific courses/tournaments
}
\examples{
getEventsForYears(2010, 2015)
}
\keyword{golf}
\keyword{pga}

