% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadShotlink.R
\name{getShotlinkTable}
\alias{getShotlinkTable}
\title{Get Shotlink tables}
\usage{
getShotlinkTable(filename)
}
\arguments{
\item{dataframe}{evnts}
}
\value{
Data Frame of shots
}
\description{
These functions load a table from the correct directory
}
\examples{
getShotlinkExtTable("shot-ext-safeway-2017.txt", safeway$local_tz)


}
\keyword{golf}
\keyword{pga}

