% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{getStationLocation}
\alias{getStationLocation}
\title{For a given station code, return the coordinates of that station}
\usage{
getStationLocation(station_code)
}
\arguments{
\item{String}{airport_code}
}
\value{
c("lat","lon")
}
\description{
For a given station code, return the coordinates of that station
}
\examples{
scrapeWeatherForTournament(masters09)
}

