% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrlResponse.R
\name{getUrlResponse}
\alias{getUrlResponse}
\title{Get response from url}
\usage{
getUrlResponse(url)
}
\arguments{
\item{string}{url}
}
\value{
string response
}
\description{
This function returns the string of a url response
}
\examples{
getUrlResponse("https://www.pga.com")
}

