% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{getWeatherObsForTournament}
\alias{getWeatherObsForTournament}
\title{For an event load all weather data for that tournament into a data frame}
\usage{
getWeatherObsForTournament(tournament)
}
\arguments{
\item{event}{from getPGAEvents() function}
}
\description{
For an event load all weather data for that tournament into a data frame
}
\examples{
getWeatherObsForTournament(masters09)
}

