% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{getWeatherObsLocationDates}
\alias{getWeatherObsLocationDates}
\title{Get Weather for location and dates}
\usage{
getWeatherObsLocationDates(loc, dates)
}
\arguments{
\item{list}{loc}

\item{list}{dates}
}
\value{
dataframe of weather observations
}
\description{
For an event load all weather data for that tournament into a data frame
}
\examples{
getWeatherObsLocationDates(safeway[,c("hole_lat", "hole_lon")], safeway[,c("start", "end")])
}

