% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historicalWeather.R
\name{historicalWeatherForEvent}
\alias{historicalWeatherForEvent}
\title{Get historical weather info for tournament}
\usage{
historicalWeatherForEvent(event)
}
\arguments{
\item{event}{event from pga}
}
\value{
saved info
}
\description{
This function finds and saves historical weather info for tournaments
}
\examples{
historicalWeatherForEvent(safeway)
}

