% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proviz.R
\name{main_scrapeProviz}
\alias{main_scrapeProviz}
\title{do all proviz scraping}
\usage{
main_scrapeProviz()
}
\arguments{
\item{string}{input name of course}
}
\value{
string of file in /data/proviz directory for the corresponding file name.
}
\description{
this function scrapes all proviz courses
}

