% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherPlot.R
\name{plotMeanWindSpeedAndDriveDist}
\alias{plotMeanWindSpeedAndDriveDist}
\title{These functions plot weather nformation}
\usage{
plotMeanWindSpeedAndDriveDist(shots, obs, plotTitle = "")
}
\description{
These functions plot weather nformation
}

