% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{scrapeWeatherForTournament}
\alias{scrapeWeatherForTournament}
\title{Scrape weather info for all dates in a tournament}
\usage{
scrapeWeatherForTournament(e)
}
\arguments{
\item{event}{from getPGAEvents() function}
}
\description{
This grabs the weather response for 3 days before up to the end of a tournament
}
\examples{
scrapeWeatherForTournament(masters09)
}

