% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{scrapeWeatherForTournaments}
\alias{scrapeWeatherForTournaments}
\title{Scrape weather info for all tournaments}
\usage{
scrapeWeatherForTournaments(events)
}
\arguments{
\item{event}{from getPGAEvents() function}
}
\value{
nothing
}
\description{
This grabs the weather response for 3 days before up to the end of multiple tournaments
respecting the rate limit of the weather underground api
}
\examples{
scrapeWeatherForTournaments(events)
}

