% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotWeatherSummary.R
\name{shotWeatherSummary}
\alias{shotWeatherSummary}
\title{shot weather summary}
\usage{
shotWeatherSummary(shot_date_time, weather, for_mark = FALSE)
}
\arguments{
\item{date_time}{shot_date_time input first year of events to return, Defaults to 0 (no lower bound)}

\item{data.frame}{weather last year of events to return, Defaults to 10000 (no upper bound)}

\item{boolean}{for_mark format data for mark or no}
}
\value{
Data Frame row of summary weather info for the shot
}
\description{
This function returns a data frame row of summary weather information for a shot
}
\examples{
shotWeatherSummary(safeway.shots[1,], safeway.weather)
}
\keyword{golf}
\keyword{pga}

