% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamartino.R
\name{yamartino}
\alias{yamartino}
\title{Calculate wind mean and variance using yamartino method}
\usage{
yamartino(obs)
}
\arguments{
\item{obs}{wind direction in degrees}
}
\value{
mean wind direction according to yamartino
}
\description{
This function finds and saves historical weather info for tournaments
}
\examples{
yamartino(safeway$lastWindSpeed)
}

