% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamartino.R
\name{yamartino_std}
\alias{yamartino_std}
\title{Calculate wind variance using yamartino method}
\usage{
yamartino_std(obs)
}
\arguments{
\item{obs}{wind direction in degrees}
}
\value{
std wind direction according to yamartino
}
\description{
This function finds and saves historical weather info for tournaments
}
\examples{
yamartino(safeway$lastWindSpeed)
}

