% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_rain_regression.R
\name{yearlyReport}
\alias{yearlyReport}
\title{Generate yearly summary report of statistics for holes, hope to identify interesting ones for analysis}
\usage{
yearlyReport(shot_weather)
}
\arguments{
\item{shot_weather}{dataframe of shots and weather}
}
\value{
saved info
}
\description{
This function calculates and saves info for a tournament
}
\examples{
yearlyReport(sw)
}

