% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeatherForCourseDate.R, R/weather.R
\name{getWeatherForCourseDate}
\alias{getWeatherForCourseDate}
\title{On a given day/course get the weather observations from a local file}
\usage{
getWeatherForCourseDate(course, dateStr)

getWeatherForCourseDate(course, dateStr)
}
\arguments{
\item{string}{input name of course}

\item{string}{date of obs}

\item{string}{input name of course}

\item{string}{date of obs}
}
\value{
summary info for event

summary info for event
}
\description{
This function loads the weather json from a local directory and writes summary information for that day

This function loads the weather json from a local directory and writes summary information for that day
}
\examples{
getWeatherForCourseDate("TPC Scottsdale")
getWeatherForCourseDate("TPC Scottsdale")
}

