% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{getWeatherForTournament}
\alias{getWeatherForTournament}
\title{Get a summary of weather for an event}
\usage{
getWeatherForTournament(course)
}
\arguments{
\item{event}{from getPGAEvents() function}
}
\value{
Dataframe of summary information from event
}
\description{
This function finds summary weather data for an individual golf event
}
\examples{
getWeatherForTournament()
}

