% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{getWeatherObsLocationDates}
\alias{getWeatherObsLocationDates}
\title{Get Weather for location and dates}
\usage{
getWeatherObsLocationDates(queries)
}
\arguments{
\item{list}{queries list of queries to grab observations from in "lat,lon-date"}
}
\value{
dataframe of weather observations
}
\description{
For an event load all weather data for that tournament into a data frame
}
\examples{
getWeatherObsLocationDates(queries)
}

