% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{getWeatherObsLocationDates}
\alias{getWeatherObsLocationDates}
\title{For an event load all weather data for that tournament into a data frame}
\usage{
getWeatherObsLocationDates(loc, dates)
}
\description{
For an event load all weather data for that tournament into a data frame
}
\examples{
getWeatherObsLocationDates(safeway[,c("hole_lat", "hole_lon")], safeway[,c("start", "end")])
}

