% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{getWeatherResponseForCoordsDate}
\alias{getWeatherResponseForCoordsDate}
\title{On a given day/course get the weather observations from a local file}
\usage{
getWeatherResponseForCoordsDate(dateStr, addr)
}
\arguments{
\item{dateString}{date of obs}

\item{coords}{hase hole_lat, hole_lon}
}
\value{
summary info for event
}
\description{
This function loads the weather json from a local directory and writes summary information for that day
}
\examples{
getWeatherResponseForCoordsDate( coords, "20150203")
}

